@@ -30,11 +30,11 @@ classifiers =
         Operating System :: Microsoft :: Windows
         Operating System :: MacOS
         Programming Language :: Python :: 3
-        Programming Language :: Python :: 3.7
         Programming Language :: Python :: 3.8
         Programming Language :: Python :: 3.9
         Programming Language :: Python :: 3.10
         Programming Language :: Python :: 3.11
+        Programming Language :: Python :: 3.12
 maintainer = SpiNNakerTeam
 maintainer_email = spinnakerusers@googlegroups.com
 keywords =
@@ -42,7 +42,7 @@ keywords =
         pynn
 
 [options]
-python_requires = >=3.7, <4
+python_requires = >=3.8, <4
 packages = find:
 zip_safe = True
 include_package_data = True
@@ -51,13 +51,9 @@ install_requires =
         matplotlib
         pyparsing>=2.2.1,<3.0.0
         quantities
-        pynn >= 0.11, != 0.12.0; python_version >= '3.8'
-        neo; python_version >= '3.8'
-        lazyarray; python_version >= '3.8'
-        # python 3.7 is no longer officially supported
-        pynn >= 0.9.1, < 0.10.0; python_version == '3.7'
-        neo >= 0.5.2, < 0.10.0; python_version == '3.7'
-        lazyarray >= 0.2.9, <= 0.4.0; python_version == '3.7'
+        pynn >= 0.11, != 0.12.0
+        neo
+        lazyarray
         # below may fail for read the docs
         scipy
         csa
